Supporting different types of newlines
Only \n (Unix) newlines were supported. The use of \r (nominally old
Macintosh) is widespread, and failed on the first government-generated
CSV file that I tried. Also common are \r\n (Windows) newlines. This
will determine which type of newline is most common in the file and
default to that.